Move painting routines to KoZoomHandler.
Okay, I commit this now, as it's mostly usable now.

What I did:
- Changed the painting of borders and background to KoZoomHandler
- Changed most of the mouse interaction to KoZoomHandler
- Changed scrolling to KoZoomHandler
- Most importantly: I got rid of all hacks we had here :-)

- Additional: The recognition for displaying the comment when mouse over
  is now the whole cell. Formerly it was the red triangle only.

What I did not (yet):
- Adjusted the text size to zoom.
- Test everything in all details, but I know several places which won't work correctly again.

Please don't write bugreports for display or interaction problems yet,
they will be fixed in the next days/weeks.

Philipp

svn path=/trunk/koffice/; revision=181444

+++ kspread/kspread_canvas.h
@@ -490,9 +490,9 @@ public:
     KSpreadHBorder( QWidget *_parent, KSpreadCanvas *_canvas, KSpreadView *_view  );
 
     int markerColumn() const { return  m_iSelectionAnchor; }
-    void resizeColumn( double resize, int nb = -1, bool makeUndo=true );
-    void adjustColumn( int _col = -1, bool makeUndo=true );
-    void equalizeColumn( int resize );
+    void resizeColumn( double resize, int nb = -1, bool makeUndo = true );
+    void adjustColumn( int _col = -1, bool makeUndo = true );
+    void equalizeColumn( double resize );
 
     void updateColumns( int from, int to );
 
@@ -560,9 +560,9 @@ public:
     KSpreadVBorder( QWidget *_parent, KSpreadCanvas *_canvas, KSpreadView *_view );
 
     int markerRow() const { return  m_iSelectionAnchor; }
-    void resizeRow( double resize, int nb = -1, bool makeUndo=true );
-    void adjustRow( int _row = -1, bool makeUndo=true );
-    void equalizeRow( int resize );
+    void resizeRow( double resize, int nb = -1, bool makeUndo = true );
+    void adjustRow( int _row = -1, bool makeUndo = true );
+    void equalizeRow( double resize );
     void updateRows( int from, int to );
 
 protected:
